+++ noodles-cram/src/data_container/compression_header/preservation_map/builder.rs
@@ -1,5 +1,3 @@
-use noodles_fasta as fasta;
-
 use super::{substitution_matrix, tag_ids_dictionary, PreservationMap};
 use crate::{writer::Options, Record};
 
@@ -18,9 +16,8 @@ impl Builder {
         self.ap_data_series_delta = options.encode_alignment_start_positions_as_deltas;
     }
 
-    pub fn update(&mut self, reference_sequence: &fasta::record::Sequence, record: &Record) {
-        self.substitution_matrix_builder
-            .update(reference_sequence, record);
+    pub fn update(&mut self, record: &Record) {
+        self.substitution_matrix_builder.update(record);
         self.tag_ids_dictionary_builder.update(record);
     }